requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Deprecated==1.2.13
Django==4.1
django-cors-middleware==1.5.0
django-oauth-toolkit==2.1.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.1
psycopg2==2.9.3
pycparser==2.21
pytz==2022.4
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
wrapt==1.14.1
